Non java projects with pom.xml files slowness
Since now we are also using `mvn help:evaluate` to find out the java
version, if the project uses maven but is not a java project (i.e.:
scala, flex, etc), every call to `__jvm_main` will take A LOT of time
and it will never find the java version in the pom.

This pull fixes that by, when no java version is found in the pom,
creating an empty `.java-version`, and later checking for existent,
non-empty `.java-version`, or loading the user `.java-version`.

It is not very elegant, but should solve the problem.

Also added a test for that.